修复183手机号端错误,精简代码

+++ src/main/java/com/superman/superman/model/CollectBean.java
@@ -20,8 +20,8 @@ public class CollectBean  {
     private String  image;
     private Integer src;
     private Integer volume;
-    private Long price;
-    private Long coupon_price;
+    private Double price;
+    private Double coupon_price;
     private Long promotion_rate;
     private Long coupon;
     private Date createTime;
